UPDATE: changed i32 to i26

+++ scalar/src/lib.rs
@@ -3,7 +3,7 @@ pub fn sum(left: u8, right: u8) -> u8 {
     left + right
 }
 
-pub fn diff(left: i32, right: i32) -> i32 {
+pub fn diff(left: i16, right: i16) -> i16 {
     left - right
 }
